Allow the path where macro expansions are emitted to be customized.
This PR is an attempt to make macro expansions more hermetic w.r.t.
DebugInfo.

Since the paths to macro expansions are emitted into DebugInfo,
the default tempdir logic will produce non-hermetic/non-reproducible
DI that would likely negatively impact the debugging experience for
remote/distributed builds. This adds a
`-emit-macro-expansion-files-path=` flag that allows that location
to be explicitly pointed to a predetermined location.

Furthermore, this change ensures that `-debug-prefix-map` remappings
are applied to the macro expansion paths when they are written to
DebugInfo.