[NFC] Rename EmitTypeMetadataRefForLayout to reflect its actual job
This visitor used to directly return type metadata, but at some
point we simplified the code a lot by just doing a type rewrite
and then calling into the normal type metadata emitter.  Apparently
we didn't rename the visitor at the time, though, so it's got this
very misleading name.