GenericSpecializer: don't re-abstract a specialized function if it has a dead indirect in-argument
If an indirect in-argument is not used in the callee, dead store elimination might have removed the preceding store to the memory location.
If we would convert such an argument to a _direct_ argument, we would insert a load before the call - but this load would load an undefined value, because the store is missing.
We want to avoid this undefined behavior. Also SILMem2Reg cannot handle it.

Usually this is no problem because there are no dead-store elimination passes before the specializer in the pipeline. But in theory, it can happen.

Add a check, which only covers the common pattern of a stack location with no stores.
This is a bad hack and the real solution would be to allow re-abstraction and pass an `undef` as direct argument.
But currently undefs are not handled very well in optimizations.