Reenable compat lib in a couple tests
This patch re-enables the compat library on a couple of tests. The issue
was that the compat library was pulling in compiler-rt symbols for
determining system version. That was fixed with a little dlsym action in
188c7bd626a9ba7a8eac598591f7bdeaedc48de6. It was using the symbols to
determine if we had voucher_needs_adopt. It just took a while to
re-enable it in these tests.