Fix String.hasPrefix and String.hasSuffix when self and arg are in NFC form
This commit fixes https://github.com/apple/swift/issues/67427. When the
string and argument are both in NFC form, besides performing the
prefix/suffix check, we also have to perform a grapheme boundary check
to make sure that the prefix/suffix are not in between an extended
grapheme cluster.

In addition, it also seems to pass a test that was previously marked as
failing.